requirements.txt
langchain
uuid
pymongo
cryptography
streamlit_antd_components